Finally realized why you might want to separate a decoder from a parser:  the parser deals with managing the tokens from the lexer, but once the parser has figured that out, then the decoder is in charge of converting the tokens to Python types.  This may seem more complicated than an all-in-one parser, but this actually concentrates the logic.  I realized that I was duplicating functionality in the lang.token class and in the old decode.decode_* functions, but by separating the old combined decoder and parser into separate elements, the interface is much cleaner.
This commit starts that work.  Probably need to break out the Grammar
and Token classes into separate files to more easily manage.  Tests
are not currently passing.  Things are in flux.

Broke up the old lang.py into more manageable grammar, lexer, and token files.

Bit the bullet and built accomodations for a time with a seconds value of sixty.  It was an awful lot of work just to return a string, but I'm a better person for it.

Figured out how to modify the lexer() to allow 'return' of a token into the
generator iterator, this makes the code sooooo much cleaner, but only part way
through fixing up all of the parser functions.

So tests are not in a working state.

I created a reasonably generic for_try_except(), if I end up using that pattern
elsewhere, we may need to create a util or a functions file.
